+++ tests/v2/test_jsonrpc_waku.nim
@@ -2,15 +2,22 @@ import
   std/[unittest, options, sets, tables, os, strutils, sequtils],
   stew/shims/net as stewNet,
   json_rpc/[rpcserver, rpcclient],
+  eth/[keys, rlp], eth/common/eth_types,
   libp2p/[standard_setup, switch, multiaddress],
   libp2p/protobuf/minprotobuf,
   libp2p/stream/[bufferstream, connection],
   libp2p/crypto/crypto,
   libp2p/protocols/pubsub/pubsub,
   libp2p/protocols/pubsub/rpc/message,
+  ../../waku/v1/node/rpc/hexstrings,
   ../../waku/v2/waku_types,
   ../../waku/v2/node/wakunode2,
-  ../../waku/v2/node/jsonrpc/[jsonrpc_types,store_api,relay_api,debug_api,filter_api,admin_api],
+  ../../waku/v2/node/jsonrpc/[store_api,
+                              relay_api,
+                              debug_api,
+                              filter_api,
+                              admin_api,
+                              private_api],
   ../../waku/v2/protocol/message_notifier,
   ../../waku/v2/protocol/waku_filter,
   ../../waku/v2/protocol/waku_store/waku_store,
@@ -70,7 +77,7 @@ procSuite "Waku v2 JSON-RPC API":
       ta = initTAddress(bindIp, rpcPort)
       server = newRpcHttpServer([ta])
 
-    installRelayApiHandlers(node, server)
+    installRelayApiHandlers(node, server, newTable[string, seq[WakuMessage]]())
     server.start()
 
     let client = newRpcHttpClient()
@@ -140,7 +147,7 @@ procSuite "Waku v2 JSON-RPC API":
       server = newRpcHttpServer([ta])
     
     # Let's connect to node 3 via the API
-    installRelayApiHandlers(node3, server)
+    installRelayApiHandlers(node3, server, newTable[string, seq[WakuMessage]]())
     server.start()
 
     let client = newRpcHttpClient()
@@ -253,7 +260,7 @@ procSuite "Waku v2 JSON-RPC API":
       ta = initTAddress(bindIp, rpcPort)
       server = newRpcHttpServer([ta])
 
-    installFilterApiHandlers(node, server)
+    installFilterApiHandlers(node, server, newTable[ContentTopic, seq[WakuMessage]]())
     server.start()
 
     let client = newRpcHttpClient()
@@ -294,7 +301,7 @@ procSuite "Waku v2 JSON-RPC API":
       ta = initTAddress(bindIp, rpcPort)
       server = newRpcHttpServer([ta])
 
-    installFilterApiHandlers(node, server)
+    installFilterApiHandlers(node, server, newTable[ContentTopic, seq[WakuMessage]]())
     server.start()
     
     node.mountFilter()
@@ -409,5 +416,186 @@ procSuite "Waku v2 JSON-RPC API":
       # Check store peer
       (response.filterIt(it.protocol == WakuStoreCodec)[0]).multiaddr == constructMultiaddrStr(storePeer)
 
+    server.stop()
     server.close()
     waitfor node.stop()
+
+  asyncTest "Private API: generate asymmetric keys and encrypt/decrypt communication":
+    let
+      nodeKey1 = crypto.PrivateKey.random(Secp256k1, rng[])[]
+      node1 = WakuNode.init(nodeKey1, bindIp, Port(60000))
+      nodeKey2 = crypto.PrivateKey.random(Secp256k1, rng[])[]
+      node2 = WakuNode.init(nodeKey2, bindIp, Port(60002))
+      nodeKey3 = crypto.PrivateKey.random(Secp256k1, rng[])[]
+      node3 = WakuNode.init(nodeKey3, bindIp, Port(60003), some(extIp), some(port))
+      pubSubTopic = "polling"
+      contentTopic = ContentTopic(1)
+      payload = @[byte 9]
+      message = WakuRelayMessage(payload: payload, contentTopic: some(contentTopic))
+      topicCache = newTable[string, seq[WakuMessage]]()
+
+    await node1.start()
+    await node1.mountRelay(@[pubSubTopic])
+
+    await node2.start()
+    await node2.mountRelay(@[pubSubTopic])
+
+    await node3.start()
+    await node3.mountRelay(@[pubSubTopic])
+
+    await node1.connectToNodes(@[node2.peerInfo])
+    await node3.connectToNodes(@[node2.peerInfo])
+
+    # Setup two servers so we can see both sides of encrypted communication
+    let
+      rpcPort1 = Port(8545)
+      ta1 = initTAddress(bindIp, rpcPort1)
+      server1 = newRpcHttpServer([ta1])
+      rpcPort3 = Port(8546)
+      ta3 = initTAddress(bindIp, rpcPort3)
+      server3 = newRpcHttpServer([ta3])
+    
+    # Let's connect to nodes 1 and 3 via the API
+    installPrivateApiHandlers(node1, server1, rng, newTable[string, seq[WakuMessage]]())
+    installPrivateApiHandlers(node3, server3, rng, topicCache)
+    installRelayApiHandlers(node3, server3, topicCache)
+    server1.start()
+    server3.start()
+
+    let client1 = newRpcHttpClient()
+    await client1.connect("127.0.0.1", rpcPort1)
+
+    let client3 = newRpcHttpClient()
+    await client3.connect("127.0.0.1", rpcPort3)
+
+    # Let's get a keypair for node3
+
+    let keypair = await client3.get_waku_v2_private_v1_asymmetric_keypair()
+
+    # Now try to subscribe on node3 using API
+
+    let sub = await client3.post_waku_v2_relay_v1_subscriptions(@[pubSubTopic])
+
+    await sleepAsync(2000.millis)
+
+    check:
+      # node3 is now subscribed to pubSubTopic
+      sub
+
+    # Now publish and encrypt a message on node1 using node3's public key
+    let posted = await client1.post_waku_v2_private_v1_asymmetric_message(pubSubTopic, message, publicKey = (%keypair.pubkey).getStr())
+    check:
+      posted
+
+    await sleepAsync(2000.millis)
+
+    # Let's see if we can receive, and decrypt, this message on node3    
+    var messages = await client3.get_waku_v2_private_v1_asymmetric_messages(pubSubTopic, privateKey = (%keypair.seckey).getStr())
+
+    check:
+      messages.len == 1
+      messages[0].contentTopic.get == contentTopic
+      messages[0].payload == payload
+    
+    # Ensure that read messages are cleared from cache
+    messages = await client3.get_waku_v2_private_v1_asymmetric_messages(pubSubTopic, privateKey = (%keypair.seckey).getStr())  
+    check:
+      messages.len == 0
+
+    server1.stop()
+    server1.close()
+    server3.stop()
+    server3.close()
+    await node1.stop()
+    await node2.stop()
+    await node3.stop()
+
+  asyncTest "Private API: generate symmetric keys and encrypt/decrypt communication":
+    let
+      nodeKey1 = crypto.PrivateKey.random(Secp256k1, rng[])[]
+      node1 = WakuNode.init(nodeKey1, bindIp, Port(60000))
+      nodeKey2 = crypto.PrivateKey.random(Secp256k1, rng[])[]
+      node2 = WakuNode.init(nodeKey2, bindIp, Port(60002))
+      nodeKey3 = crypto.PrivateKey.random(Secp256k1, rng[])[]
+      node3 = WakuNode.init(nodeKey3, bindIp, Port(60003), some(extIp), some(port))
+      pubSubTopic = "polling"
+      contentTopic = ContentTopic(1)
+      payload = @[byte 9]
+      message = WakuRelayMessage(payload: payload, contentTopic: some(contentTopic))
+      topicCache = newTable[string, seq[WakuMessage]]()
+
+    await node1.start()
+    await node1.mountRelay(@[pubSubTopic])
+
+    await node2.start()
+    await node2.mountRelay(@[pubSubTopic])
+
+    await node3.start()
+    await node3.mountRelay(@[pubSubTopic])
+
+    await node1.connectToNodes(@[node2.peerInfo])
+    await node3.connectToNodes(@[node2.peerInfo])
+
+    # Setup two servers so we can see both sides of encrypted communication
+    let
+      rpcPort1 = Port(8545)
+      ta1 = initTAddress(bindIp, rpcPort1)
+      server1 = newRpcHttpServer([ta1])
+      rpcPort3 = Port(8546)
+      ta3 = initTAddress(bindIp, rpcPort3)
+      server3 = newRpcHttpServer([ta3])
+    
+    # Let's connect to nodes 1 and 3 via the API
+    installPrivateApiHandlers(node1, server1, rng, newTable[string, seq[WakuMessage]]())
+    installPrivateApiHandlers(node3, server3, rng, topicCache)
+    installRelayApiHandlers(node3, server3, topicCache)
+    server1.start()
+    server3.start()
+
+    let client1 = newRpcHttpClient()
+    await client1.connect("127.0.0.1", rpcPort1)
+
+    let client3 = newRpcHttpClient()
+    await client3.connect("127.0.0.1", rpcPort3)
+
+    # Let's get a symkey for node3
+
+    let symkey = await client3.get_waku_v2_private_v1_symmetric_key()
+
+    # Now try to subscribe on node3 using API
+
+    let sub = await client3.post_waku_v2_relay_v1_subscriptions(@[pubSubTopic])
+
+    await sleepAsync(2000.millis)
+
+    check:
+      # node3 is now subscribed to pubSubTopic
+      sub
+
+    # Now publish and encrypt a message on node1 using node3's symkey
+    let posted = await client1.post_waku_v2_private_v1_symmetric_message(pubSubTopic, message, symkey = (%symkey).getStr())
+    check:
+      posted
+
+    await sleepAsync(2000.millis)
+
+    # Let's see if we can receive, and decrypt, this message on node3    
+    var messages = await client3.get_waku_v2_private_v1_symmetric_messages(pubSubTopic, symkey = (%symkey).getStr())
+
+    check:
+      messages.len == 1
+      messages[0].contentTopic.get == contentTopic
+      messages[0].payload == payload
+    
+    # Ensure that read messages are cleared from cache
+    messages = await client3.get_waku_v2_private_v1_symmetric_messages(pubSubTopic, symkey = (%symkey).getStr())
+    check:
+      messages.len == 0
+
+    server1.stop()
+    server1.close()
+    server3.stop()
+    server3.close()
+    await node1.stop()
+    await node2.stop()
+    await node3.stop()